// src/lib/server/supabase/types.ts
export type Database = {
  public: {
    Tables: {
      perfume_inventory: {
        Row: {
          created_at: string
          id: number
          is_tester: boolean | null
          perfume_id: number | null
          price: number
          url: string | null
          volume: number
          website: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          is_tester?: boolean | null
          perfume_id?: number | null
          price: number
          url?: string | null
          volume: number
          website?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          is_tester?: boolean | null
          perfume_id?: number | null
          price?: number
          url?: string | null
          volume?: number
          website?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "perfume_inventory_perfume_id_fkey"
            columns: ["perfume_id"]
            isOneToOne: false
            referencedRelation: "perfumes"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "perfume_inventory_website_fkey"
            columns: ["website"]
            isOneToOne: false
            referencedRelation: "websites"
            referencedColumns: ["name"]
          },
        ]
      }
      perfumes: {
        Row: {
          concentration: string | null
          created_at: string
          fragrantica_url: string | null
          gender: string | null
          house: string
          id: number
          image_url: string | null
          name: string
        }
        Insert: {
          concentration?: string | null
          created_at?: string
          fragrantica_url?: string | null
          gender?: string | null
          house: string
          id?: number
          image_url?: string | null
          name: string
        }
        Update: {
          concentration?: string | null
          created_at?: string
          fragrantica_url?: string | null
          gender?: string | null
          house?: string
          id?: number
          image_url?: string | null
          name?: string
        }
        Relationships: []
      }
      websites: {
        Row: {
          created_at: string
          id: number
          logo_url: string | null
          name: string
          url: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          logo_url?: string | null
          name: string
          url?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          logo_url?: string | null
          name?: string
          url?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      get_distinct_houses:
        | {
            Args: Record<PropertyKey, never>
            Returns: {
              house: string
            }[]
          }
        | {
            Args: {
              search_query?: string
            }
            Returns: {
              house: string
            }[]
          }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never